requirements.txt
# Runtime deps
Pillow==7.1.*
reportlab==3.5.*

# code maint
autoflake
isort
black
twine

# tests
pytest
mypy
